pyproject.toml
[project]
name = "astrobro"
version = "0.9.0"
description = "your pocket astrologer 💫"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["natal[report]>=0.9.3", "streamlit-shortcuts"]

[dependency-groups]
dev = ["ipykernel>=6.29.5", "pytest>=8.3.4", "watchdog>=6.0.0"]

[tool.ruff.lint.isort]
no-sections = true
